d_a_lstair: Match leftover moveBG (demo partial)

+++ src/d/actor/d_a_lstair.cpp
@@ -245,7 +245,7 @@ void daLStair_c::moveBG() {
 
     if (mAppearTimer < 0) {
         if (cLib_calcTimer(&mTimer) == 0) {
-            int r0 = 60;
+            s16 r0 = 60;
             cLib_addCalc(&mStairYOffset, target, 1.0f / r0, maxStep, minStep);
         }
     }
